chore: Refactor build node connections (#9803)
Another refactoring! The plus side of all these is more readable code (hopefully), but they're mostly to ease the loki integration. This one renames a lot of the vars/functions in `build-node-connections` so they're consistent with the actual graphql types and field class names as per https://graphql.org/graphql-js/type/#graphqlobjecttype

Another benefit of this refactor is the ability to create types (connections in this case) one by one instead of having to do them as a whole batch. This will be eventually needed for incremental builds

Update: I slipped a refactor to `build-node-types` in as well.